third_party/blink/web_tests/wpt_internal/overscroll-gestures/snap-area-ignores-overscroll.html

<!DOCTYPE html>
<html >
<title>Overscroll position: out of flow positioned</title>
<meta name="author" content="github.com/vmpstr">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
body, html {
  margin: 0;
  padding: 0;
}

#overscroll {
  top: -500px;
  width: 100%;
  height: 10px;
  background: red;
}
#container {
  width: 500px;
  height: 500px;
  overflow: scroll;
  scroll-snap-type: y mandatory;
}
.content {
  scroll-snap-align: start;
  height: 50px;
  &.top {
    background: green;
  }
  &.bottom {
    background: yellow;
  }
}
.spacer {
  height: 500px;
  border: 1px solid black;
}
</style>

<div id=container overscrollcontainer>
  <div id=overscroll></div>
  <div id=target class="top content"></div>
  <div class=spacer></div>
  <div class="bottom content"></div>
  <div class=spacer></div>
</div>

<button command="toggle-overscroll" commandfor="overscroll"></button>

<script>
test(() => {
  assert_equals(target.getBoundingClientRect().y, 0);
}, "snapped position in an overscrollcontainer scroller");
</script>
